backend/requirements.txt
fastapi[standard]
websockets
pyjwt
passlib[bcrypt]
sqlalchemy
sqlmodel
python-dotenv